services/bot/requirements.txt
# Telegram Bot requirements
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
aiogram>=3.3.0
kafka-python>=2.0.2
aiohttp>=3.9.0

